+++ client/pages/Assessment/Assessment.tsx
@@ -1,19 +1,52 @@
 import './Assessment.scss'
 import React, { useEffect } from 'react'
-import Navigation from '@client/components/Navigation'
-import { Route, Switch } from 'react-router-dom'
-import { useNavigationVisible } from '@client/store/ui/navigation'
-import { BasePaths } from '@client/basePaths'
-import OriginalDataPoint from '@client/pages/OriginalDataPoint'
-import { useCountryIso } from '@client/hooks'
-import { useAppDispatch } from '@client/store'
-import { AssessmentActions } from '@client/store/assessment'
 import { useParams } from 'react-router'
+import { Route, Switch } from 'react-router-dom'
+
 import { AssessmentName } from '@meta/assessment'
 import { Areas } from '@meta/area'
+
+import { useAppDispatch } from '@client/store'
+import { AssessmentActions } from '@client/store/assessment'
+import { useNavigationVisible } from '@client/store/ui/navigation'
+import { AssessmentSectionActions } from '@client/store/pages/assessmentSection'
+import { useCountryIso } from '@client/hooks'
+import { BasePaths } from '@client/basePaths'
+
+import Navigation from '@client/components/Navigation'
+import MessageCenter from '@client/components/MessageCenter'
+import OriginalDataPoint from '@client/pages/OriginalDataPoint'
 import AssessmentSection from '@client/pages/AssessmentSection'
 import DataExport from '@client/pages/DataExport'
-import MessageCenter from '@client/components/MessageCenter'
+
+const SectionWrapper: React.FC = (props) => {
+  const { children } = props
+
+  const dispatch = useAppDispatch()
+  const countryIso = useCountryIso()
+  const { assessmentName, cycleName, section } = useParams<{
+    assessmentName: AssessmentName
+    cycleName: string
+    section: string
+  }>()
+
+  useEffect(() => {
+    dispatch(
+      AssessmentSectionActions.getTableSections({
+        assessmentName,
+        cycleName,
+        section,
+        countryIso,
+      })
+    )
+
+    return () => {
+      dispatch(AssessmentSectionActions.reset())
+    }
+  }, [countryIso, assessmentName, cycleName, section])
+
+  return <>{React.Children.toArray(children)}</>
+}
 
 const Assessment: React.FC = () => {
   const navigationVisible = useNavigationVisible()
@@ -21,10 +54,7 @@ const Assessment: React.FC = () => {
   const countryIso = useCountryIso()
   const isDataExport = countryIso && !Areas.isISOCountry(countryIso)
 
-  const { assessmentName, cycleName } = useParams<{
-    assessmentName: AssessmentName
-    cycleName: string
-  }>()
+  const { assessmentName, cycleName } = useParams<{ assessmentName: AssessmentName; cycleName: string }>()
 
   useEffect(() => {
     dispatch(
@@ -44,9 +74,20 @@ const Assessment: React.FC = () => {
       <Switch>
         {/* <Route path={BasePaths.assessmentHome} component={AssessmentHome} /> */}
         {/* <Route path={BasePaths.assessmentDataDownload} component={AssessmentDataDownload} /> */}
-        <Route exact path={BasePaths.Assessment.section()} component={isDataExport ? DataExport : AssessmentSection} />
+        <Route
+          exact
+          path={BasePaths.Assessment.section()}
+          render={() => <SectionWrapper>{isDataExport ? <DataExport /> : <AssessmentSection />}</SectionWrapper>}
+        />
         {/* <Route exact path={[`${BasePaths.odp}:odpId/`, BasePaths.odp]} component={OriginalDataPoint} /> */}
-        <Route path={BasePaths.Assessment.OriginalDataPoint.section()} component={OriginalDataPoint} />
+        <Route
+          path={BasePaths.Assessment.OriginalDataPoint.section()}
+          render={() => (
+            <SectionWrapper>
+              <OriginalDataPoint />
+            </SectionWrapper>
+          )}
+        />
       </Switch>
     </div>
   )
