+++ client/pages/DataExport/DataExport.tsx
@@ -1,14 +1,8 @@
 import './DataExport.scss'
-import React, { useEffect } from 'react'
-import { useParams } from 'react-router'
+import React from 'react'
 
 // import { Objects } from '@core/utils'
 // import { useDataExportCountries, useDataExportSelection } from '@webapp/store/page/dataExport'
-import { AssessmentName } from '@meta/assessment'
-import { AssessmentSectionActions } from '@client/store/pages/assessmentSection'
-import { useAppDispatch } from '@client/store'
-import { useCountryIso } from '@client/hooks'
-import { useAssessmentSection } from '@client/store/assessment'
 
 // import CountrySelect from './CountrySelect'
 // import VariableSelect from './VariableSelect'
@@ -16,31 +10,6 @@ import { useAssessmentSection } from '@client/store/assessment'
 // import ResultsTable from './ResultsTable'
 
 const DataExport: React.FC = () => {
-  const assessmentSection = useAssessmentSection()
-
-  const { assessmentName, cycleName, section } = useParams<{
-    assessmentName: AssessmentName
-    cycleName: string
-    section: string
-  }>()
-
-  const dispatch = useAppDispatch()
-  const countryIso = useCountryIso()
-
-  useEffect(() => {
-    dispatch(AssessmentSectionActions.reset())
-    dispatch(
-      AssessmentSectionActions.getTableSections({
-        assessmentName,
-        cycleName,
-        section,
-        countryIso,
-      })
-    )
-  }, [countryIso, assessmentSection])
-
-  if (!assessmentSection) return null
-
   // const assessmentSection = useParamSection()
   // const countries = useDataExportCountries()
   // const selection = useDataExportSelection(assessmentSection)
